+++ architecture/available_arch.json
@@ -43,6 +43,18 @@
     "guide": "",
     "available": 1
   },
+    {
+    "name": "Z80",
+    "alias": ["Z80"],
+    "file": "Z80",
+    "img": "/img/z80_logo.png",
+    "alt": "Z80",
+    "id": "select_confZ80",
+    "examples": [],
+    "description": "The Z80 is an 8-bit microprocessor that was designed by Zilog and introduced in 1976. It became very popular in the late 1970s and early 1980s, especially in home computers and embedded systems.",
+    "guide": "",
+    "available": 1
+  },
   {
     "name": "New Architecture",
     "alias": [],
