+++ src/core/core.mjs
@@ -58,31 +58,18 @@ import { resetStats } from "./executor/stats.mts";
 
 // Conditional import for the WASM compiler based on the environment (web or Deno)
 
-import wasm_web_init, {
-    Color as Color_web,
-    ArchitectureJS as ArchitectureJS_web,
-} from "./compiler/web/creator_compiler.js";
-import {
-    Color as Color_deno,
-    ArchitectureJS as ArchitectureJS_deno,
-} from "./compiler/deno/creator_compiler.js";
+// import {
+//     Color as Color_deno,
+//     ArchitectureJS as ArchitectureJS_deno,
+// } from "./compiler/creatorCompiler/deno/creator_compiler.js";
 
-let Color;
-let ArchitectureJS;
-if (isDeno) {
-    // Deno HAS to be imported like this, as it doesn't provide a default
-    Color = Color_deno;
-    ArchitectureJS = ArchitectureJS_deno;
-} else if (isWeb) {
-    Color = Color_web;
-    ArchitectureJS = ArchitectureJS_web;
-    // in the web, we MUST call the default
-    wasm_web_init();
-} else {
-    throw new Error(
-        "Unsupported environment: neither Deno nor web browser detected",
-    );
-}
+// let Color;
+// let ArchitectureJS;
+// if (isDeno) {
+//     // Deno HAS to be imported like this, as it doesn't provide a default
+//     Color = Color_deno;
+//     ArchitectureJS = ArchitectureJS_deno;
+// }
 
 export let code_assembly = "";
 export let update_binary = "";
@@ -981,8 +968,8 @@ function prepareArchitecture(
         return Math.max(max, instruction.nwords || 1);
     }, 1);
 
-    // Convert to JSON for WASM
-    const architectureJson = JSON.stringify(architectureObj);
+    // // Convert to JSON for WASM
+    // const architectureJson = JSON.stringify(architectureObj);
 
     // Dump the architecture JSON to a file for debugging
     if (dump) {
@@ -996,11 +983,6 @@ function prepareArchitecture(
                 `Could not write architecture file: ${writeError.message}`,
             );
         }
-    }
-
-    // Initialize WASM compiler if not skipped
-    if (!skipCompiler) {
-        arch = ArchitectureJS.from_json(architectureJson);
     }
 
     return architectureObj;
@@ -1168,11 +1150,10 @@ export function load_library(lib_str) {
 
 // compilation
 
-export function assembly_compile(code, enable_color, compiler = "default") {
-    const ret = assembly_compiler(
+export async function assembly_compile(code, compiler) {
+    const ret = await assembly_compiler(
         code,
         false,
-        enable_color ? Color.Ansi : Color.Off,
         compiler,
     );
     switch (ret.status) {
