+++ src/web/components/SimulatorView.vue
@@ -31,10 +31,10 @@ import Calculator from "./simulator/Calculator.vue"
 import Stats from "./simulator/Stats.vue"
 
 import { architecture } from "@/core/core.mjs"
-import { instructions } from "@/core/compiler/compiler.mjs"
 
 export default {
   props: {
+    instructions: { type: Array, required: true },
     browser: String,
     os: { type: String, required: true },
     dark: { type: Boolean, required: true },
@@ -70,7 +70,6 @@ export default {
   data() {
     return {
       architecture,
-      instructions,
 
       // stack
       callee_subrutine: "",
