--- z80.o
@@ -1,32 +1,0 @@
-{
-  "instructions_binary": [
-    {
-      "Break": null,
-      "Address": "0x0",
-      "Label": "main",
-      "loaded": "0x00",
-      "user": null,
-      "_rowVariant": "",
-      "visible": true,
-      "globl": true
-    },
-    {
-      "Break": null,
-      "Address": "0x2",
-      "Label": "",    
-      "loaded": "0x80",
-      "user": null,
-      "_rowVariant": "",
-      "visible": true
-    }
-  ],
-  "instructions_tag": [
-    {
-      "tag": "main",
-      "addr": 0,
-      "globl": true
-    }
-  ],
-  "data_section": [
-  ]
-}
